用户删除： 用户作为子库文件上传者时为假删除，作为任务相关负责人等或子任务相关负责人审核人等或子库文件审核相关人员等时为真删除

+++ src/main/java/com/rengu/cosimulation/service/UserService.java
@@ -3,10 +3,8 @@ package com.rengu.cosimulation.service;
 import com.rengu.cosimulation.entity.*;
 import com.rengu.cosimulation.enums.ResultCode;
 import com.rengu.cosimulation.exception.ResultException;
-import com.rengu.cosimulation.repository.ProjectRepository;
-import com.rengu.cosimulation.repository.RoleRepository;
-import com.rengu.cosimulation.repository.SubtaskRepository;
-import com.rengu.cosimulation.repository.UserRepository;
+import com.rengu.cosimulation.repository.*;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.CachePut;
@@ -25,20 +23,23 @@ import java.util.*;
  * Date: 2019/2/12 16:15
  */
 @Service
+@Slf4j
 public class UserService implements UserDetailsService {
     private final UserRepository userRepository;
     private final RoleService roleService;
     private final DepartmentService departmentService;
     private final ProjectRepository projectRepository;
     private final SubtaskRepository subtaskRepository;
+    private final SublibraryFilesRepository sublibraryFilesRepository;
 
     @Autowired
-    public UserService(UserRepository userRepository, RoleService roleService, DepartmentService departmentService, ProjectRepository projectRepository, SubtaskRepository subtaskRepository) {
+    public UserService(UserRepository userRepository, RoleService roleService, DepartmentService departmentService, ProjectRepository projectRepository, SubtaskRepository subtaskRepository, SublibraryFilesRepository sublibraryFilesRepository) {
         this.userRepository = userRepository;
         this.roleService = roleService;
         this.departmentService = departmentService;
         this.projectRepository = projectRepository;
         this.subtaskRepository = subtaskRepository;
+        this.sublibraryFilesRepository = sublibraryFilesRepository;
     }
 
     // 保存用户 , 一个角色
@@ -79,14 +80,9 @@ public class UserService implements UserDetailsService {
         return userRepository.save(users);
     }
 
-    //保存管理员用户
-    /*public Users saveAdminUser(Users users, String roleName) {
-        return saveUser(users);
-    }*/
-
     // 查询所有用户
     public List<Users> getAll() {
-        return userRepository.findAll();
+        return userRepository.findByDeleted(false);
     }
 
     // 根据用户名查询用户是否存在
@@ -94,7 +90,7 @@ public class UserService implements UserDetailsService {
         if (StringUtils.isEmpty(username)) {
             return false;
         }
-        return userRepository.existsByUsername(username);
+        return userRepository.existsByUsernameAndDeleted(username, false);
     }
 
     // 根据用户名查询用户
@@ -103,7 +99,7 @@ public class UserService implements UserDetailsService {
         if (!hasUserByUsername(username)) {
             throw new ResultException(ResultCode.USER_USERNAME_NOT_FOUND_ERROR);
         }
-        return userRepository.findByUsername(username).get();
+        return userRepository.findByUsernameAndDeleted(username, false).get();
     }
 
     @Override
@@ -156,21 +152,47 @@ public class UserService implements UserDetailsService {
             throw new ResultException(ResultCode.USER_ID_NOT_FOUND_ERROR);
         }
         Users users = getUserById(userId);
-        List<Project> projectList = projectRepository.findByPicOrCreator(users, users);
-        if(projectList.size() > 0){
-            for(Project project : projectList){
-                if(project.getPic().getId().equals(userId)){
-                    project.setPic(null);
-                }else {
-                    project.setCreator(null);
+        List<SubDepotFile> subDepotFiles = sublibraryFilesRepository.findByUsers(users);
+        if(subDepotFiles.size() > 0){             // 用户为子库文件上传者时则假删除，否则直接删除
+            users.setEnabled(false);
+            users.setDeleted(true);
+            userRepository.save(users);
+        }else {
+            List<Project> projectList = projectRepository.findByPicOrCreator(users, users);
+            if(projectList.size() > 0){
+                for(Project project : projectList){
+                    if(project.getPic().getId().equals(userId)){
+                        project.setPic(null);
+                    }else {
+                        project.setCreator(null);
+                    }
                 }
             }
+            List<Subtask> subtaskList = subtaskRepository.findByUsersOrProofSetContainingOrAuditSetContainingOrCountSetContainingOrApproveSetContaining(users, users, users, users, users);
+            if(subtaskList.size() > 0){
+                for(Subtask subtask : subtaskList){
+                    if(subtask.getUsers().getId().equals(userId)){
+                        subtask.setUsers(null);
+                    }
+                    subtask.getProofSet().remove(users);
+                    subtask.getAuditSet().remove(users);
+                    subtask.getCountSet().remove(users);
+                    subtask.getApproveSet().remove(users);
+                }
+                subtaskRepository.saveAll(subtaskList);
+            }
+            List<SubDepotFile> subDepotFileList = sublibraryFilesRepository.findByProofSetContainingOrAuditSetContainingOrCountSetContainingOrApproveSetContaining(users, users, users, users);
+            if(subDepotFileList.size() > 0){
+                for(SubDepotFile subDepotFile : subDepotFileList){
+                    subDepotFile.getProofSet().remove(users);
+                    subDepotFile.getAuditSet().remove(users);
+                    subDepotFile.getCountSet().remove(users);
+                    subDepotFile.getApproveSet().remove(users);
+                }
+                sublibraryFilesRepository.saveAll(subDepotFileList);
+            }
+            userRepository.delete(users);
         }
-        List<Subtask> subtaskList = subtaskRepository.findByUsersOrProofSetContainingOrAuditSetContainingOrCountSetContainingOrApproveSetContaining(users, users, users, users, users);
-        if(subtaskList.size() > 0){
-            subtaskRepository.deleteInBatch(subtaskList);
-        }
-        userRepository.delete(users);
         return users;
     }
